+++ src/devtools/components/Mutations/Mutations.js
@@ -13,7 +13,14 @@ import Warning from "../Images/Warning";
 import "../WatchedQueries/WatchedQueries.less";
 
 const mutationLabel = (mutationId, mutation) => {
-  const mutationName = getOperationName(mutation.mutation);
+  const mutationName = getOperationName(
+    // Apollo Client >= v2.5 includes `mutation.mutationString`. Versions prior
+    // include `mutation.mutation` which is an AST represetation of the
+    // mutation.
+    mutation.mutationString
+      ? parse(mutation.mutationString)
+      : mutation.mutation,
+  );
 
   if (!mutationName) {
     return mutationId;
@@ -184,7 +191,10 @@ class WatchedMutation extends React.Component {
       mutation.metadata.component.displayName;
     const displayName = componentDisplayName || reactComponentDisplayName;
 
-    const mutationString = print(mutation.mutation);
+    // Apollo Client >= v2.5 includes `mutation.mutationString`. Versions prior
+    // include `mutation.mutation` which is an AST represetation of the
+    // mutation.
+    const mutationString = mutation.mutationString || print(mutation.mutation);
 
     return (
       <div className={classnames("main", { loading: mutation.loading })}>